More compiler fixes, CSV coding improvement 1) Fixed a UINT compiler warning 2) Initialized Running Avg variable for CSV generation 3) Another attempt at making Copy Image routine work 4) Removed unused resource

+++ PitsideConsole/ArtLib/ArtNet.cpp
@@ -19,7 +19,7 @@ int TimeoutRead(SOCKET s, char* buf, int cbBuf, int flags, int timeout, bool* pf
       }
       else
       {
-        if(timeGetTime() - tmNow > timeout)
+        if(timeGetTime() - tmNow > (UINT)timeout)
         {
           // timed out
           *pfConnectionLost = true;
